+++ CHANGELOG.md
@@ -25,6 +25,7 @@ CHANGELOG
 - `bots.output.amqptopic`:
   - The default exchange must not be declared (#1295).
   - Unencodable characters are prepended by backslashes by default. Otherwise Unicode characters can't be encoded and sent (#1296).
+  - Gracefully close AMQP connection on shutdown of bot.
 
 ### Documentation
 - Bots: document redis cache parameters.